+++ projects/form-validation/src/lib/models/form-validation-configuration.ts
@@ -9,6 +9,7 @@ export const enum SupportedFrameworks {
 
 export interface FormValidationConfig {
   invalidFeedbackTemplate?: Type<BaseInvalidFeedbackComponent>;
+  defaultErrorMessages: {[key: string]: string};
   invalidClasses: string;
   skipValidation: boolean;
   targetSelector: string;