+++ projects/form-validation/CHANGELOG.md
@@ -1,6 +1,20 @@
-## 0.0.3 (2022-10-05)
+## 0.0.5 (2022-10-06)
 
 ### Features
 
-- Fix bug `Can't find package`
+- Add default error messages to support default Validator from `@angular/form`
+
+## 0.0.4 (2022-10-05)
+
+### Fix Bugs
+
+- Not show error message when submit form
+
+## 0.0.3 (2022-10-05)
+
+### Fix bugs
+
+- `Can't find package`
+
+### Docs
 - Add Stackblitz demo link to README